refactor: pathExtension 추출 로직을 Request로 이동
StatusLine보다는 Request에서 직접 해주는게 더 적절하다고 판단.
