+++ src/main/java/webserver/http/Request.java
@@ -22,6 +22,22 @@ public class Request {
     }
 
     public String pathExtension() {
-        return requestMessage.pathExtension();
+        String path = path();
+
+        String[] splitPath = path.split("/");
+
+        if (splitPath.length == 0) {
+            return "";
+        }
+
+        String[] splitDot = splitPath[splitPath.length - 1].split("\\.");
+
+        String extension = "";
+
+        if (1 < splitDot.length) {
+            extension = splitPath[splitPath.length - 1].split("\\.")[1];
+        }
+
+        return extension;
     }
 }
