+++ src/main/java/webserver/http/statusline/RequestStatusLine.java
@@ -50,24 +50,4 @@ public class RequestStatusLine extends StatusLine {
                                     .add(protocol())
                                     .toString();
     }
-
-    public String pathExtension() {
-        String path = path();
-
-        String[] splitPath = path.split("/");
-
-        if (splitPath.length == 0) {
-            return "";
-        }
-
-        String[] splitDot = splitPath[splitPath.length - 1].split("\\.");
-
-        String extension = "";
-
-        if (1 < splitDot.length) {
-            extension = splitPath[splitPath.length - 1].split("\\.")[1];
-        }
-
-        return extension;
-    }
 }
