+++ src/test/java/webserver/http/statusline/RequestStatusLineTest.java
@@ -126,98 +126,4 @@ class RequestStatusLineTest {
                 )
         );
     }
-
-    @ParameterizedTest
-    @MethodSource("pathExtension")
-    void pathExtension(String desc, RequestStatusLine requestStatusLine, String expectedExtension) {
-        assertThat(requestStatusLine.pathExtension())
-                .as("status line에서 path의 확장자 가져오기 : %s", desc)
-                .isEqualTo(expectedExtension);
-    }
-
-    static Stream<Arguments> pathExtension() {
-        return Stream.of(
-                Arguments.of(
-                        "확장자가 있는 path",
-                        new RequestStatusLine(
-                                new HashMap<String, String>() {{
-                                    put("path", "/user/create.html");
-                                    put("method", "GET");
-                                    put("protocolVersion", "HTTP/1.1");
-                                }}
-                        ),
-                        "html"
-                ), Arguments.of(
-                        "확장자가 있는 path /로 끝남",
-                        new RequestStatusLine(
-                                new HashMap<String, String>() {{
-                                    put("path", "/user/create.html/");
-                                    put("method", "GET");
-                                    put("protocolVersion", "HTTP/1.1");
-                                }}
-                        ),
-                        "html"
-                ), Arguments.of(
-                        "확장자와 쿼리스트링이 함께 있는 path",
-                        new RequestStatusLine(
-                                new HashMap<String, String>() {{
-                                    put("path", "/user/create.html?userId=javajigi&password=password&name=%EB%B0%95%EC%9E%AC%EC%84%B1&email=javajigi%40slipp.net");
-                                    put("method", "GET");
-                                    put("protocolVersion", "HTTP/1.1");
-                                }}
-                        ),
-                        "html"
-                ),Arguments.of(
-                        "확장자와 쿼리스트링이 함께 있는 path /로 끝남",
-                        new RequestStatusLine(
-                                new HashMap<String, String>() {{
-                                    put("path", "/user/create.html?userId=javajigi&password=password&name=%EB%B0%95%EC%9E%AC%EC%84%B1&email=javajigi%40slipp.net/");
-                                    put("method", "GET");
-                                    put("protocolVersion", "HTTP/1.1");
-                                }}
-                        ),
-                        "html"
-                ), Arguments.of(
-                        "확장자가 없는 path",
-                        new RequestStatusLine(
-                                new HashMap<String, String>() {{
-                                    put("path", "/user/create");
-                                    put("method", "GET");
-                                    put("protocolVersion", "HTTP/1.1");
-                                }}
-                        ),
-                        ""
-                ), Arguments.of(
-                        "root만 있는 path",
-                        new RequestStatusLine(
-                                new HashMap<String, String>() {{
-                                    put("path", "/");
-                                    put("method", "GET");
-                                    put("protocolVersion", "HTTP/1.1");
-                                }}
-                        ),
-                        ""
-                ),Arguments.of(
-                        "비어 있는 path",
-                        new RequestStatusLine(
-                                new HashMap<String, String>() {{
-                                    put("path", "");
-                                    put("method", "GET");
-                                    put("protocolVersion", "HTTP/1.1");
-                                }}
-                        ),
-                        ""
-                ), Arguments.of(
-                        "확장자가 없는데 쿼리스트링이 있는 path",
-                        new RequestStatusLine(
-                                new HashMap<String, String>() {{
-                                    put("path", "/user/create?userId=javajigi&password=password&name=%EB%B0%95%EC%9E%AC%EC%84%B1&email=javajigi%40slipp.net");
-                                    put("method", "GET");
-                                    put("protocolVersion", "HTTP/1.1");
-                                }}
-                        ),
-                        ""
-                )
-        );
-    }
 }
